third_party/blink/web_tests/external/wpt/html/semantics/forms/resetting-a-form/reset-form-event-realm-expected.txt

This is a testharness.js-based test.
[FAIL] reset()'s event must be fired in the Realm of the target
  assert_equals: the event must be created in the realm of the target expected function "function Function() { [native code] }" but got function "function Function() { [native code] }"
Harness: the test ran to completion.
